% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{build_layer_plot}
\alias{build_layer_plot}
\title{Function to build the jittered layer plot + dendrogram. From library(tasic2018analysis)}
\usage{
build_layer_plot(anno, dend, cocl, cluster_ids, seed_val = 42,
  right_pad = 10, fillColor = c("#ECE09C", "#F7F2DA", "#A7D7DF",
  "#C1E5E7", "#D4EDED", "#B3E3E3"), textSize = 2, maxPerCluster = Inf,
  seed = 1)
}
\value{
a ggplot2 plot object
}
\description{
Function to build the jittered layer plot + dendrogram. From library(tasic2018analysis)
}
