% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{build_legend_plot}
\alias{build_legend_plot}
\title{This function builds a legend plot.  THIS FUNCTION IS NOW IN library(scrattch.vis)}
\usage{
build_legend_plot(data_source, genes = c("Hspa8", "Snap25", "Gad2",
  "Vip"), autorange = "auto", minrange = 0, maxrange = 10,
  scale_type = "scale.log", pfontsize = 14)
}
\value{
a ggplot2 plot object
}
\description{
This function builds a legend plot.  THIS FUNCTION IS NOW IN library(scrattch.vis)
}
