% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{calc_beta}
\alias{calc_beta}
\title{Calculates the beta score}
\usage{
calc_beta(y, spec.exp = 2)
}
\arguments{
\item{y}{a numeric vector (typically corresponding to cluster proportions)}

\item{spec.exp}{should be left at default value of 2}
}
\value{
beta score
}
\description{
This score is a marker score that is a combination of specificity and sparsity
}
