% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusterInfoMTG}
\alias{clusterInfoMTG}
\title{Cluster information for MTG data set.}
\format{A data frame with 75 rows and 11 variables:
\describe{
  \item{cluster_id}{cluster identifier}
  \item{cluster_label}{cluster name}
  \item{cluster_color}{cluster color in MTG paper}
  \item{cluster_type_label}{What broad class is that cluster}
  \item{all others}{can be ignored}
}}
\source{
\url{https://www.biorxiv.org/content/early/2018/08/05/384826}
}
\usage{
clusterInfoMTG
}
\description{
Cluster information (including name, color, and type) for MTG data sets.
}
\keyword{datasets}
