% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{errorBarPlot}
\alias{errorBarPlot}
\title{Plots and error bar}
\usage{
errorBarPlot(vals, sampleType, col = standardColors(), legend = TRUE,
  elwd = 2, ylim = NA, xlim = NA, length = 0.1, ...)
}
\description{
Plots and error bar
}
