% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FI_layer5_count_data}
\alias{FI_layer5_count_data}
\title{snRNA-seq data from frontoinsula}
\format{A data frame with 30370 rows (genes) and 879 variables (nuclei):
\describe{
  \item{each column}{gene expression levels for all genes in that nucleus}
}}
\source{
\url{https://github.com/AllenInstitute/VENcelltypes}
}
\usage{
FI_layer5_count_data
}
\description{
CPM values (rounded to the nearest integer) for all 879 nuclei passing QC in 
  the human FI data set.
}
\keyword{datasets}
