% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FI_layer5_sample_information}
\alias{FI_layer5_sample_information}
\title{Sample information for nuclei from frontoinsula.}
\format{A data frame with 879 rows (nuclei) and 11 variables:
\describe{
  \item{cluster_id}{OLD cluster identifier}
  \item{cluster_label}{OLD cluster name}
  \item{cluster_color}{OLD cluster color}
  \item{patient_label}{donor of origin}
  \item{genecountsGTzeroFPKM_label}{Genes with expression > 0}
  \item{mappedReads_label}{Number of mapped reads per nucleus}
  \item{PCT_INTRONIC_BASES_label}{Percent of base pairs mapping to introns}
  \item{PCT_HUMAN_EXONS_BASES_label}{Percent of base pairs mapping to exons}
  \item{all others}{are not used in manuscript, but have reasonable column headers}
}}
\source{
\url{https://github.com/AllenInstitute/VENcelltypes}
}
\usage{
FI_layer5_sample_information
}
\description{
Sample information and various metadata for all 879 nuclei passing QC in 
  the human FI data set.
}
\keyword{datasets}
