% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{findFromGroups}
\alias{findFromGroups}
\title{Feturns the summary gene expression value across a group}
\usage{
findFromGroups(datExpr, groupVector, fn = "mean")
}
\arguments{
\item{datExpr}{a matrix of data (rows=genes, columns=samples)}

\item{groupVector}{character vector corresponding to the group (e.g., cell type)}

\item{fn}{Summary function to use (default = mean)}
}
\value{
Summary matrix (genes x groups)
}
\description{
Feturns the summary gene expression value across a group
}
