% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_functions.R
\name{FindNeighbors.Seurat}
\alias{FindNeighbors.Seurat}
\title{(From Seurat) FindNeighbors (required for Windows compatibility)}
\usage{
\method{FindNeighbors}{Seurat}(object, reduction = "pca", dims = 1:10,
  assay = NULL, features = NULL, k.param = 20, compute.SNN = TRUE,
  prune.SNN = 1/15, nn.eps = 0, verbose = TRUE,
  force.recalc = FALSE, do.plot = FALSE, graph.name = NULL, ...)
}
\arguments{
\item{reduction}{Reduction to use as input for building the SNN}

\item{dims}{Dimensions of reduction to use as input}

\item{assay}{Assay to use in construction of SNN}

\item{features}{Features to use as input for building the SNN}

\item{do.plot}{Plot SNN graph on tSNE coordinates}

\item{graph.name}{Optional naming parameter for stored SNN graph. Default is
assay.name_snn.}
}
\description{
(From Seurat) FindNeighbors (required for Windows compatibility)
}
