% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{formatPlotGenes}
\alias{formatPlotGenes}
\title{Function for getting the plotting genes (to avoid typing it out a bunch of times)}
\usage{
formatPlotGenes(plotGenes)
}
\value{
gene vector
}
\description{
Function for getting the plotting genes (to avoid typing it out a bunch of times)
}
