% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_functions.R
\name{get_feather_data}
\alias{get_feather_data}
\title{Read data from a directory of feather files. From library(scrattch.io)}
\usage{
get_feather_data(feather_dir, genes, group_by, group_ids)
}
\arguments{
\item{feather_dir}{Directory containing feather files.}

\item{genes}{Genes to read}

\item{group_by}{desc_bases to use for grouping samples}

\item{group_ids}{ID values to use for filtering}
}
\description{
Read data from a directory of feather files. From library(scrattch.io)
}
