% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{get_node_dend}
\alias{get_node_dend}
\title{Function to subset a dendrogram by retrieving a node with a given attribute.}
\usage{
get_node_dend(x, match_attr, match_value)
}
\value{
dendrogram subset
}
\description{
Function to subset a dendrogram by retrieving a node with a given attribute.
}
