% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{getAnovaPvalforApply}
\alias{getAnovaPvalforApply}
\title{ANOVA for use with the apply function}
\usage{
getAnovaPvalforApply(x, varLabels, varWeights = NULL)
}
\description{
ANOVA for use with the apply function
}
