% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{getBetaScore}
\alias{getBetaScore}
\title{Wrapper for calc_beta}
\usage{
getBetaScore(propExpr, returnScore = TRUE, spec.exp = 2)
}
\arguments{
\item{propExpr}{proportions of cells in a given cluster with CPM/FPKM > 1 (or 0, HCT uses 1)}

\item{returnScore}{TRUE returns scores, FALSE results ranks}

\item{spec.exp}{do not change from default (2)}
}
\value{
a vector of beta scores
}
\description{
Wrapper for calc_beta
}
