% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{getTopMarkersByProp2}
\alias{getTopMarkersByProp2}
\title{Function for getting top marker genes}
\usage{
getTopMarkersByProp2(propExpr, n = 1, excludeGenes = NULL,
  medianExpr = NA, fcThresh = 0.5, minProp = 0)
}
\description{
Function for getting top marker genes
}
