% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{getTopMarkersByPropNew}
\alias{getTopMarkersByPropNew}
\title{Function for getting top marker genes}
\usage{
getTopMarkersByPropNew(propExpr, medianExpr, propDiff = 0,
  propMin = 0.5, medianFC = 1, excludeGenes = NULL,
  sortByMedian = TRUE)
}
\description{
Function for getting top marker genes
}
