% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_functions.R
\name{LogSeuratCommand}
\alias{LogSeuratCommand}
\title{(From Seurat) logs a command}
\usage{
LogSeuratCommand(object, return.command = FALSE)
}
\arguments{
\item{object}{Name of Seurat object}
}
\value{
returns the Seurat object with command stored
}
\description{
Logs a command run, storing the name, timestamp, and argument list. Stores in
the Seurat object
}
