% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mito_genes}
\alias{mito_genes}
\title{Mitochondrial-associated genes.}
\format{A character vector with 1182 elements:
\describe{
  \item{mito_genes}{cluster identifier}
}}
\source{
\url{https://www.biorxiv.org/content/early/2018/08/05/384826}
}
\usage{
mito_genes
}
\description{
Vector of mitochondrial-associated genes
}
\keyword{datasets}
