% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{mouse2human2}
\alias{mouse2human2}
\title{Conversion between mouse and human gene names (not used)}
\usage{
mouse2human2(mouse, m2h)
}
\description{
Conversion between mouse and human gene names (not used)
}
