% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{plot_dend2}
\alias{plot_dend2}
\title{Plots the dendrogram in a convenient format.  From  library (scrattch.hicat)}
\usage{
plot_dend2(dend, dendro_data = NULL, node_size = 1, r = NULL,
  label_color = NULL, main = "", rMin = -0.6)
}
\description{
Plots the dendrogram in a convenient format.  From  library (scrattch.hicat)
}
