% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_functions.R
\name{qcPlot}
\alias{qcPlot}
\title{Makes QC plots.  From library(patchseqtools)}
\usage{
qcPlot(anno, name, scaleLimits = c(-5000, 12000), scaleBreaks = seq(0,
  12000, 2000), scaleLabels = seq(0, 12, 2), ylab = "value",
  yVal = 0, width = 7, height = 3, fileName = gsub("\\\\.", "_",
  gsub("_label", "", name)), screenPlot = TRUE)
}
\arguments{
\item{anno}{Annotation data frame for patch-seq}

\item{scaleLimits}{scaleLimits}

\item{scaleBreaks}{scaleBreaks}

\item{scaleLabels}{scaleLabels}

\item{ylab}{ylab}

\item{fileName}{fileName}

\item{dendcluster_anno}{what to cluster by}

\item{groupvars}{what to group by}
}
\value{
the plot is returned
}
\description{
Makes QC plots.  From library(patchseqtools)
}
