% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{rankCompleteDendPlot}
\alias{rankCompleteDendPlot}
\title{Wrapper for plot_dend}
\usage{
rankCompleteDendPlot(input = NULL, l.rank = NULL, dend = NULL,
  label_color = NULL, node_size = 3, main = "Tree",
  distFun = function(x) return(dist(1 - (1 + cor(x))/2)), ...)
}
\description{
Wrapper for plot_dend
}
