% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{renameClusters}
\alias{renameClusters}
\title{Another function for renaming clusters based on data and meta-data}
\usage{
renameClusters(sampleInfo, clusterInfo, propExpr, medianExpr,
  propDiff = 0, propMin = 0.5, medianFC = 1, excludeGenes = NULL,
  majorGenes = c("GAD1", "SLC17A7", "SLC1A3"),
  majorLabels = majorGenes, broadGenes = majorGenes, propLayer = 0.3,
  layerNameColumn = "layer_label", getLayer = function(x)
  return(as.numeric(substr(as.character(x), 2, 2))))
}
\description{
See renameAndOrderClusters in library(scrattch.hicat)
}
