% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{reorder.dend}
\alias{reorder.dend}
\title{Reorders a clustering tree.  Function from library(scrattch.hicat)}
\usage{
\method{reorder}{dend}(dend, l.rank, verbose = FALSE)
}
\description{
Reorders a clustering tree.  Function from library(scrattch.hicat)
}
