% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{sample_heatmap_plot}
\alias{sample_heatmap_plot}
\title{Heatmaps of gene expression for individual samples.   THIS FUNCTION IS NOW IN library(scrattch.vis)}
\usage{
sample_heatmap_plot(data_source, genes = c("Hspa8", "Snap25", "Gad2",
  "Vip"), group_by = "cluster", groups = 1:10, top_labels = "layer",
  sample_n = 0, scale_mode = "scale.log", showall = F,
  autorange = "auto", minrange = 0, maxrange = 10, pfontsize = 14,
  expand = F, rotatelabels = F, showlines = F, showids = T)
}
\arguments{
\item{genes}{A character vector containing gene symbols to be plotted.}

\item{data}{A data frame containing gene expression values. The first column should be sample_name}

\item{anno}{Sample annotations. The first column should be sample_name, and each annotation should have \_id, \_label, and \_color columns}

\item{grouping}{A character vector specifying the desc base that should be used to group cells}

\item{group_order}{Optional: Explicit specification of group order by supplying a vector of group_ids.}

\item{log_scale}{Logical , determines if data is log scaled before plotting. Default = FALSE.}

\item{normalize_rows}{Logical, determines if heatmaps will be normalized for each gene. Default = FALSE.}

\item{font_size}{numeric object, the font size (in pts) used to make the plot.}

\item{label_height}{numeric object, Percent of the plot height that should be used for the labels (0 to 100). Default is 25.}

\item{label_type}{Label shape, "angle" or "square"}

\item{max_width}{numeric object, percent of plot width that should be used for maximum expression values (0 to 100). Default is 10.}

\item{return_type}{What values to return - can be "plot", "data", or "both". Default is "plot".}
}
\value{
a ggplot2 plot object
}
\description{
Heatmaps of gene expression for individual samples.   THIS FUNCTION IS NOW IN library(scrattch.vis)
}
