% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sex_genes}
\alias{sex_genes}
\title{Sex-associated genes.}
\format{A character vector with 2433 elements:
\describe{
  \item{sex_genes}{cluster identifier}
}}
\source{
\url{https://www.biorxiv.org/content/early/2018/08/05/384826}
}
\usage{
sex_genes
}
\description{
Vector of sex-associated genes
}
\keyword{datasets}
