% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{subsampleCells}
\alias{subsampleCells}
\title{Subsample cells.  THIS FUNCTION IS IN library(mfishtools)}
\usage{
subsampleCells(clustersF, subSamp = 25, seed = 5)
}
\arguments{
\item{subSamp}{maximum number of values for each category to subsample}

\item{seed}{for reproducibility}

\item{clusters}{vector of cluster labels (or any category) in factor or character format}
}
\value{
returns a vector of TRUE / FALSE with a maximum of subSamp TRUE calls per category
}
\description{
Subsets a categorical vector to include up to a maximum number of values for each category.
}
