% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{t.test.l.paired}
\alias{t.test.l.paired}
\title{Paired t.test wrapper for use with the "apply" function}
\usage{
\method{t}{test.l.paired}(x)
}
\description{
Paired t.test wrapper for use with the "apply" function
}
