% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{updateAndOrderClusters}
\alias{updateAndOrderClusters}
\title{Function for updating and ordering clusters based on meta-data}
\usage{
updateAndOrderClusters(sampleInfo,
  classNameColumn = "cluster_type_label",
  layerNameColumn = "layer_label", matchNameColumn = "cellmap_label",
  regionNameColumn = "Region_label",
  newColorNameColumn = "cellmap_color", otherColumns = NULL,
  topGenes = NULL, classLevels = c("inh", "exc", "glia"),
  getLayer = function(x) return(as.numeric(substr(as.character(x), 2,
  2))), sep = "_")
}
\description{
See renameAndOrderClusters in library(scrattch.hicat)
}
